sftp to s3 example/requirements.txt
# imports
protobuf==3.8.0
six==1.12.0
tensorflow-estimator==2.2.0
azure-functions
grpcio==1.26.0
azure-functions-worker
azure-storage-blob
paramiko
boto3